Load parameters by name
